+++ js/data.js
@@ -1,3 +1,4 @@
+
 let now = new Date();
 now.setMinutes(now.getMinutes() - now.getTimezoneOffset());
 document.getElementById('to-date').value = now.toISOString().slice(0, 16);