+++ CHANGELOG.md
@@ -14,6 +14,15 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2021.2.1.26] - 2021-04-26
+
+### Changed
+- font Milliard
+
+### Added
+- power consumption overview chart
+- charm for displaying debug information
+
 ## [2021.2.1.14] - 2021-04-14
 
 ### Added